refactor: apply changed common response format
