+++ src/main/java/app/bottlenote/review/controller/ReviewController.java
@@ -26,7 +26,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static app.bottlenote.global.data.response.GlobalResponse.success;
 import static app.bottlenote.user.exception.UserExceptionCode.REQUIRED_USER_ID;
 import static app.bottlenote.user.exception.UserExceptionCode.USER_NOT_FOUND;
 
@@ -39,57 +38,37 @@ public class ReviewController {
 	private final ReviewService reviewService;
 
 	@PostMapping
-	public ResponseEntity<GlobalResponse> createReview(@RequestBody @Valid ReviewCreateRequest reviewCreateRequest) {
+	public ResponseEntity<?> createReview(@RequestBody @Valid ReviewCreateRequest reviewCreateRequest) {
 
 		Long currentUserId = SecurityContextUtil.getUserIdByContext().
 			orElseThrow(() -> new UserException(USER_NOT_FOUND));
 
 		ReviewCreateResponse createdReview = reviewService.createReview(reviewCreateRequest, currentUserId);
-		return ResponseEntity.ok(success(createdReview));
+		return GlobalResponse.ok(createdReview);
 	}
 
 	@GetMapping("/{alcoholId}")
-	public ResponseEntity<GlobalResponse> getReviews(@PathVariable Long alcoholId, @ModelAttribute ReviewPageableRequest reviewPageableRequest) {
-
+	public ResponseEntity<?> getReviews(@PathVariable Long alcoholId, @ModelAttribute ReviewPageableRequest reviewPageableRequest) {
 		Long currentUserId = SecurityContextUtil.getUserIdByContext().orElse(-1L);
-
-		log.info("currentUserId is : {} \nPageable INFO  : {}", currentUserId, reviewPageableRequest.toString());
-
 		PageResponse<ReviewListResponse> pageResponse = reviewService.getReviews(
 			alcoholId,
-                reviewPageableRequest,
+			reviewPageableRequest,
 			currentUserId
 		);
 
-		return ResponseEntity.ok(
-			success(pageResponse.content(), MetaService.createMetaInfo().add("pageable", pageResponse.cursorPageable())
-			));
+		return GlobalResponse.ok(pageResponse.content(), MetaService.createMetaInfo().add("pageable", pageResponse.cursorPageable()));
 	}
 
-	/**
-	 * 위스키의 리뷰 상세를 조회하는 API 입니다.
-	 * <p>
-	 * 유저 아이디가 존재하지않을때 userId를 -1L 로 조회 :
-	 * "isPicked": false 값으로만 조회됩니다.
-	 * "isMyReview": false 값으로만 조회됩니다.
-	 * "isLikedByMe": false 값으로만 조회됩니다.
-	 * "hasReplyByMe": false 값으로만 조회됩니다.
-	 *
-	 * @param reviewId
-	 * @return
-	 */
 	@GetMapping("/detail/{reviewId}")
-	public ResponseEntity<GlobalResponse> getDetailReview(@PathVariable Long reviewId) {
+	public ResponseEntity<?> getDetailReview(@PathVariable Long reviewId) {
 
 		Long currentUserId = SecurityContextUtil.getUserIdByContext().orElse(-1L);
 
-		return ResponseEntity.ok(
-			success(reviewService.getDetailReview(reviewId, currentUserId))
-		);
+		return GlobalResponse.ok(reviewService.getDetailReview(reviewId, currentUserId));
 	}
 
 	@GetMapping("/me/{alcoholId}")
-	public ResponseEntity<GlobalResponse> getMyReviews(@ModelAttribute ReviewPageableRequest reviewPageableRequest, @PathVariable Long alcoholId) {
+	public ResponseEntity<?> getMyReviews(@ModelAttribute ReviewPageableRequest reviewPageableRequest, @PathVariable Long alcoholId) {
 
 		Long currentUserId = SecurityContextUtil.getUserIdByContext().orElseThrow(
 			() -> new UserException(REQUIRED_USER_ID)
@@ -97,21 +76,17 @@ public class ReviewController {
 
 		PageResponse<ReviewListResponse> myReviews = reviewService.getMyReviews(reviewPageableRequest, alcoholId, currentUserId);
 
-		return ResponseEntity.ok(
-			success(myReviews.content(), MetaService.createMetaInfo().add("pageable", myReviews.cursorPageable()))
-		);
+		return GlobalResponse.ok(myReviews.content(), MetaService.createMetaInfo().add("pageable", myReviews.cursorPageable()));
 	}
 
 	@PatchMapping("/{reviewId}")
-	public ResponseEntity<GlobalResponse> modifyReview(@RequestBody @Valid ReviewModifyRequest reviewModifyRequest, @PathVariable Long reviewId) {
+	public ResponseEntity<?> modifyReview(@RequestBody @Valid ReviewModifyRequest reviewModifyRequest, @PathVariable Long reviewId) {
 
 		Long currentUserId = SecurityContextUtil.getUserIdByContext().orElseThrow(
 			() -> new UserException(REQUIRED_USER_ID)
 		);
 
-		return ResponseEntity.ok(
-			success(reviewService.modifyReview(reviewModifyRequest, reviewId, currentUserId))
-		);
+		return GlobalResponse.ok(reviewService.modifyReview(reviewModifyRequest, reviewId, currentUserId));
 	}
 
 	@PatchMapping("/{reviewId}/display")
@@ -119,7 +94,6 @@ public class ReviewController {
 		@PathVariable Long reviewId,
 		@Valid @RequestBody ReviewStatusChangeRequest status
 	) {
-
 		Long currentUserId = SecurityContextUtil.getUserIdByContext().orElseThrow(
 			() -> new UserException(REQUIRED_USER_ID)
 		);
@@ -128,14 +102,11 @@ public class ReviewController {
 	}
 
 	@DeleteMapping("/{reviewId}")
-	public ResponseEntity<GlobalResponse> deleteReview(@PathVariable Long reviewId) {
-
+	public ResponseEntity<?> deleteReview(@PathVariable Long reviewId) {
 		Long currentUserId = SecurityContextUtil.getUserIdByContext().orElseThrow(
 			() -> new UserException(REQUIRED_USER_ID)
 		);
 
-		return ResponseEntity.ok(
-			success(reviewService.deleteReview(reviewId, currentUserId))
-		);
+		return GlobalResponse.ok(reviewService.deleteReview(reviewId, currentUserId));
 	}
 }
